Review environment variables for SpanLimits
Some of the variables that the code was using were non-standard,
and some of the limits had defaults when the spec specified none.
In preparation for the SDK documentation (which will include a
section on variables that can be used to configure it) these were
spotted and fixed.

This change also includes changes to the validation of those values.
The code was croaking, but this has been relaxed to a warning and
making the code react as if the value had not been set. This is in
line with the validation for the sampler argument, which also comes
from the environment.

The Console exporter tests had to be touched because the SpanLimits
was mistakenly assigning a default limit to the event attribute
length, and the code in OpenTelemetry::Attributes was reacting to
this by calling `substr` on the value, which was therefore
"stringified". When Data::Dumper saw that value, it was printed as a
string. Since this is no longer the case, the test failed.

This highlights that the Console exporter tests are excessivey
brittle. It would be better if they deserialised the exported string
and compared the data structure, but that will be left for a future
change.