Use bigfloat instead of bignum in ratio sampler
This is a minor change to sidestep an issue caused at a distance
by loading this pragma. The effect can be seen in the following
snippets:

    $ perl -Mbignum -E '
        use Math::BigFloat;
        say Math::BigFloat->new(42)->ffround;
    '
    Can't locate object method "ffround" via package "Math::BigInt" at -e line 1.

    $ perl -Mbigfloat -E '
        use Math::BigFloat;
        say Math::BigFloat->new(42)->ffround;
    '
    42

The issue (reported in [GH #16]) means that loading
OpenTelemetry::SDK (and indirectly loading `bignum`) would
at-a-distance break any code that was attempting to (directly or
indirectly) call `ffround` on a Math::BigFloat object.

Although this is most definitely not an OpenTelemetry::SDK bug,
it's unclear whose bug it _is_, and unclear how easy it would be
to fix upstream, so this solution which avoids the problem is
probably the best short-term approach.

Since `bifloat` was only added in Math::BigInt 0.65, we are
reporting this as a new dependency, to ensure it is available.

[GH #16]: https://github.com/jjatria/perl-opentelemetry-sdk/issues/16