Die on errors during import
The [specification] states that

> OpenTelemetry implementations MUST NOT throw unhandled exceptions
> at run time.

but it also says that

> The [...] SDK MAY _fail fast_ and cause the application to fail on
> initialization, e.g. because of a bad user config or environment,
> but MUST NOT cause the application to fail later at run time, e.g.
> due to dynamic config settings received from the Collector.

Before this change, the SDK was not "failing fast" for the sake of
application stability. However, this could result in setups that were
misconfigured but kept running, giving the user the incorrect impression
that things had gone according to plan. This made debugging harder.

The situation was made worse because before 56d6689b66cec4f8d1950750ed044f872e561ce1
(which added the OTLP exporter as a dependency of the SDK) because in
that case the SDK was defaulting to a broken configuration state, and
failing silently. This is no longer the case, so any remaining cases of
a misconfigured SDK are bugs that should be addressed, and are probably
going unnoticed by users.

Fixes #18, but at what cost?

[specification]: https://opentelemetry.io/docs/specs/otel/error-handling/#basic-error-handling-principles